Move rattle code into public folder
I don't know if this is necessarily the best decision. Problem: compiled
react code uses relative paths, but you are not allowed to have anything
other than astro/html files inside the pages directory.

Since the code is compiled and minimized already, it is quite difficult
to move only the assets to public, and then update the paths.