Remove tailwind config
These extended commands are actually not being called anywhere. In
addition, this is outdated tailwind v2. Lastly, all the tailwind in the
project is called as a script in their respective files, rather than
actually using this config (the index.astro uses bootstrap, not
tailwind). So, this is safe to remove.